refactor(workflow): adjust `Store` interface to use `Put` method
- Moves the `Assert` functions into a new `spec` package
- Adjusts the `Store` interface. Removes `Create` and `Update` in favor of a `Put` function.
- Adds a `Workflow.Validate` function (used in `store.Put`) that ensures a valid Workflow (has ID, DatasetID, OwnerID, Created fields)
- Adjusts `orchestrator.CreateWorkflow` to use `store.Put`
- Adds `json.Marshaller` and `json.Unmarshaller` interface to `Hook` and `Trigger` interfaces, and adds `MarshalJSON` and `UnmarshalJSON` methods to `TestTrigger` and `TestHook`
