+++ automation/workflow/trigger_test.go
@@ -1,12 +1,13 @@
 package workflow
 
+import "encoding/json"
+
 // A TestTrigger implements the Trigger interface & keeps track of the number
 // of times it had been advanced
 type TestTrigger struct {
-	workflowID   ID
-	enabled      bool
-	triggerType  TriggerType
-	AdvanceCount int
+	enabled      bool        `json:"enabled"`
+	triggerType  TriggerType `json:"type"`
+	AdvanceCount int         `json:"advanceCount"`
 }
 
 var _ Trigger = (*TestTrigger)(nil)
@@ -15,9 +16,8 @@ var _ Trigger = (*TestTrigger)(nil)
 var TestTriggerType = TriggerType("Test Trigger")
 
 // NewTestTrigger returns an enabled `TestTrigger`
-func NewTestTrigger(wid ID) *TestTrigger {
+func NewTestTrigger() *TestTrigger {
 	return &TestTrigger{
-		workflowID:   wid,
 		enabled:      true,
 		triggerType:  TestTriggerType,
 		AdvanceCount: 0,
@@ -28,7 +28,7 @@ func (tt *TestTrigger) Enabled() bool {
 	return tt.enabled
 }
 
-func (tt *TestTrigger) SetEnable(enabled bool) error {
+func (tt *TestTrigger) SetEnabled(enabled bool) error {
 	tt.enabled = enabled
 	return nil
 }
@@ -42,6 +42,13 @@ func (tt *TestTrigger) Advance() error {
 	return nil
 }
 
-func (tt *TestTrigger) WorkflowID() ID {
-	return tt.workflowID
+func (tt *TestTrigger) MarshalJSON() ([]byte, error) {
+	return json.Marshal(tt)
+}
+
+func (tt *TestTrigger) UnmarshalJSON(d []byte) error {
+	if tt == nil {
+		tt = &TestTrigger{}
+	}
+	return json.Unmarshal(d, tt)
 }
